GeminiLegion-kapa/requirements.txt
fastapi
uvicorn[standard]
google-generativeai>=0.7.0 # For system_instruction and latest features
python-dotenv
httpx # For Minions to make external HTTP calls
pydantic>=2.0 # Often comes with FastAPI, but good to specify